+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 - add ApiRestController
 - change method signature of .route by using named parameters instead of map
 
+## 0.1.13
+- Adding helfer functions to route-controller for parsing body-data like post-values
+
+## 0.1.9 (2015-05-03)
+Feature:
+- add RouteError for transporting http-status-codes and messages to this layer; to handle error output not for your own
+
+## 0.1.7 (2015-05-03)
+Feature:
+- add async/await handling
+
 ## 0.1.6 (2015-04-24)
 Feature:
 - RouteControllers execute-methode gets the request as parameter